[SPT2-171] refactor(add-product): added useAddProduct hook to handle the mutation
